+++ test/libsolidity/syntaxTests/nameAndTypeResolution/578_private_member_getter_call_without_parentheses.sol
@@ -1,0 +1,17 @@
+contract A{
+    function f() public pure{
+
+    }
+}
+contract B{
+    A private a;
+}
+contract C{
+    B b;
+    function f() public view{
+        b.a.f();
+    }
+}
+
+// ----
+// TypeError: (141-144): Member "a" not found or not visible after argument-dependent lookup in contract B.